pyproject.toml
[project]
name = "coremind"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
  "langgraph>=0.1.20",
  "langchain>=0.2.0",
  "langchain-core>=0.2.0",
  "langchain-community>=0.2.0",
  "langchain-groq>=0.1.0",
  "langchain-ollama>=0.1.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.5.0",
  "typing-extensions>=4.9.0",
  "rich>=13.7.0",
  "loguru>=0.7.2",
  "google-auth>=2.29.0",
  "google-auth-oauthlib>=1.2.0",
  "google-api-python-client>=2.120.0",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0.0",
  "ruff>=0.6.8",
  "pre-commit>=3.8.0",
]
